rufft
Module itertools

Signal processing functions which operate on Iterable implemenentors. Named itertools after the itertool crate

Modules

complex

Functions

pad

Clone and pad the real valued input collection with the floating point type F to the length len

pad_to_nearest_power_of_two

Clone and pad the real valued input collection with floating point value F to the nearest power of two length

zero_pad

Clone and zero pad the real valued input collection to the length len

zero_pad_to_nearest_power_of_two

Clone and zero pad the real valued input collection to the nearest power of two length