rufft::itertools
Module complex

Functions

pad_clone

Clone and pad the complex valued input collection with the complex floating point type F by the length n. Meaning the output collection will have a final length of x.len() + n

pad_inplace

Pad the complex valued input collection inplace with the complex floating point type Complex<F> by the length n. Meaning the collection will have a final length of x.len() + n

pad_to_nearest_power_of_two_clone

Clone and pad the complex valued input collection with complex floating point value F to the nearest power of two length

pad_to_nearest_power_of_two_inplace

Pad the complex valued input collection in place with the complex floating point value Complex<F> to the nearest power of two length

zero_pad_clone

Clone and zero pad the complex valued input collection by the length n. Meaning the output collection will have a final length of x.len() + n

zero_pad_inplace

Zero pad the complex valued input collection by the length n. Meaning the collection will have a final length of x.len() + n

zero_pad_to_nearest_power_of_two_clone

Clone and zero pad the complex valued input collection to the nearest power of two length

zero_pad_to_nearest_power_of_two_inplace

Zero pad the real valued input collection inplace to the nearest power of two length