rufft
Module traits

Traits

ComplexFloatIterable

ComplexFloatIterableMut

ExtendableIterable

Extendable iterable trait that can be implemented on dynamic collect types to push values onto the collection

Fft

Trait containing fft method which computes the fft of the real valued collection type and returns a complex value collection. The return collection type does not need to be the same type as the type the trait is implemented on

FloatIterable

FloatIterable2d

FloatIterable3d

FloatIterableMut

Iterable

Iterable trait to encapsulate collection types which have a length, are reversible, and are iterable