rufft::traits
Trait Iterable

pub trait Iterable: FromIterator<Self::OwnedItem>
where
    for<'c> Self::Item<'c>: Deref<Target = Self::OwnedItem>,
    for<'c> Self: 'c + Clone,
{
    type OwnedItem;
    type Item<'collection>
       where Self: 'collection;
    type MutItem<'collection>: DerefMut<Target = Self::OwnedItem>
       where Self: 'collection;
    type Iterator<'collection>: ExactSizeIterator<Item = Self::Item<'collection>> + DoubleEndedIterator<Item = Self::Item<'collection>>
       where Self: 'collection;
    type MutIterator<'collection>: ExactSizeIterator<Item = Self::MutItem<'collection>> + DoubleEndedIterator<Item = Self::MutItem<'collection>>
       where Self: 'collection;

    // Required methods
    fn iter<'c>(&'c self) -> Self::Iterator<'c>;
    fn iter_mut<'c>(&'c mut self) -> Self::MutIterator<'c>;

    // Provided method
    fn len(&self) -> usize { ... }
}

Iterable trait to encapsulate collection types which have a length, are reversible, and are iterable

Required Associated Types

type OwnedItem

The owned collection item. For example f64 for Vec<f64>

type Item<'collection> where Self: 'collection

The item produced by the collection’s iterator. For example &f64 for Vec<f64>

type MutItem<'collection>: DerefMut<Target = Self::OwnedItem> where Self: 'collection

type Iterator<'collection>: ExactSizeIterator<Item = Self::Item<'collection>> + DoubleEndedIterator<Item = Self::Item<'collection>> where Self: 'collection

The iterator produced by the collection

type MutIterator<'collection>: ExactSizeIterator<Item = Self::MutItem<'collection>> + DoubleEndedIterator<Item = Self::MutItem<'collection>> where Self: 'collection

Required Methods

fn iter<'c>(&'c self) -> Self::Iterator<'c>

Create an iterator from the collection over Self::Item types

fn iter_mut<'c>(&'c mut self) -> Self::MutIterator<'c>

Create a mutable iterator from the collection over Self::MutItem types

Provided Methods

fn len(&self) -> usize

Return the length of the collection. Default implementation provided based on the fact that an ExactSizeIterator implementation is required but may be more performant to override the implementation

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.