rufft::traits
Trait Fft

pub trait Fft<F: Float + FloatConst + NumAssign + 'static>
where
    for<'c> Self: Iterable<OwnedItem = F, Item<'c> = &'c F, MutItem<'c> = &'c mut F>,
    usize: AsPrimitive<F>,
{
    // Provided method
    fn fft<C>(&self) -> C
       where for<'c> C: Iterable<OwnedItem = Complex<F>, Item<'c> = &'c Complex<F>, MutItem<'c> = &'c mut Complex<F>> + IndexMut<usize, Output = Complex<F>>,
             usize: AsPrimitive<F> { ... }
}

Trait containing fft method which computes the fft of the real valued collection type and returns a complex value collection. The return collection type does not need to be the same type as the type the trait is implemented on

Provided Methods

fn fft<C>(&self) -> C
where for<'c> C: Iterable<OwnedItem = Complex<F>, Item<'c> = &'c Complex<F>, MutItem<'c> = &'c mut Complex<F>> + IndexMut<usize, Output = Complex<F>>, usize: AsPrimitive<F>,

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.